requirements.txt
fastapi[all]
uvicorn[standard]
joblib==1.1.0
pandas
scikit_learn==1.1.1
streamlit==1.10.0
SQLAlchemy==1.4.35
psycopg2-binary
python-dotenv